feat: refactor nextjs based news parser
